+++ app/Http/Controllers/Room/RoomTypeController.php
@@ -5,24 +5,39 @@ namespace App\Http\Controllers\Room;
 use App\Models\RoomType;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\RoomTypeRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class RoomTypeController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
-    {
-        $roomTypes = RoomType::all();
-        return view('admin.roomtype.index',compact('roomTypes'));
-    }
+
+     use AuthorizesRequests; 
+
+
+
+     public function index()
+     {
+         if (Gate::denies('viewAny', RoomType::class)) {
+             abort(403, 'Unauthorized');
+         }
+ 
+         $roomTypes = RoomType::all();
+         return view('admin.roomtype.index', compact('roomTypes'));
+     }
 
     /**
      * Show the form for creating a new resource.
      */
     public function create()
     {
+        if (Gate::denies('viewAny', RoomType::class)) {
+            abort(403, 'Unauthorized');
+        }
+
         return view('admin.roomtype.create');
     }
 
@@ -69,7 +84,10 @@ class RoomTypeController extends Controller
      */
     public function destroy(string $id)
     {
+
         $roomType = RoomType::findOrFail($id);
+        $this->authorize('delete',  $roomType);
+
         $roomType->delete();
         return redirect()->route('roomtype.index')->with('success', 'Room type deleted successfully!');
     }
